Conflicts:
	src/server/game/DataStores/DBCStores.cpp
	src/server/game/DataStores/DBCStores.h
	src/server/game/DataStores/DBCfmt.h
	src/server/game/Entities/Unit/Unit.cpp
	src/server/game/Entities/Unit/Unit.h
	src/server/game/Handlers/AuctionHouseHandler.cpp
	src/server/game/Handlers/GroupHandler.cpp
	src/server/game/Movement/Spline/MovementTypedefs.h
	src/server/game/Spells/Auras/SpellAuraEffects.cpp
	src/server/game/Spells/SpellInfo.cpp
	src/server/scripts/EasternKingdoms/ScarletEnclave/chapter1.cpp
	src/server/scripts/Spells/spell_generic.cpp

+++ src/server/game/AuctionHouse/AuctionHouseMgr.cpp
@@ -660,12 +660,12 @@ void AuctionEntry::SaveToDB(SQLTransaction& trans) const
     stmt->setUInt32(1, auctioneer);
     stmt->setUInt32(2, itemGUIDLow);
     stmt->setUInt32(3, owner);
-    stmt->setInt32 (4, int32(buyout));
-    stmt->setUInt64(5, uint64(expire_time));
+    stmt->setUInt32(4, buyout);
+    stmt->setUInt32(5, uint32(expire_time));
     stmt->setUInt32(6, bidder);
-    stmt->setInt32 (7, int32(bid));
-    stmt->setInt32 (8, int32(startbid));
-    stmt->setInt32 (9, int32(deposit));
+    stmt->setUInt32(7, bid);
+    stmt->setUInt32(8, startbid);
+    stmt->setUInt32(9, deposit);
     trans->Append(stmt);
 }
